+++ src/jvmMain/kotlin/org/kbson/internal/ext/getCurrentTime.kt
@@ -13,10 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kbson.ext
+package org.kbson.internal.ext
 
-import kotlin.system.getTimeMillis
-
-actual fun getCurrentTimeInMillis(): Long = getTimeMillis()
+actual fun getCurrentTimeInMillis(): Long = System.currentTimeMillis()
 
 actual fun getCurrentTimeInSeconds(): Int = (getCurrentTimeInMillis() / 1000).toInt()
