+++ src/nativeMain/kotlin/org/kbson/internal/ext/AtomicInt.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kbson.ext
+package org.kbson.internal.ext
 
 import kotlin.native.concurrent.AtomicInt
